Make a new structure in a project  and refactor code
