+++ src/App.js
@@ -1,0 +1,16 @@
+import React, { Component } from 'react';
+import './App.css';
+import ToDoList from './containers/ToDoList/ToDoList';
+
+
+class App extends Component {
+  render() {
+    return (
+      <div>
+        <ToDoList />
+      </div>
+    )
+  }
+}
+
+export default App;
