+++ src/components/NewTodoForm/NewTodoForm.js
@@ -1,0 +1,10 @@
+import React from 'react';
+
+const NewTodoForm = ({onChange, draft, onSubmit}) => (
+    <div>
+        <input type="text" onChange={onChange} value={draft}/>
+        <button onClick={onSubmit}> Add </button>
+    </div>
+)
+
+export default NewTodoForm;
